Add tests to check JUnit3 compatibility
This is intended to define a base line of behavior that future
modifications must satisfy.

Adds a RunOutputRule that provides support for testing test
running code. It provides builders for defining expectations in
terms of their output to System.out, intercepts the output and
checks that it matches the expectations.

Adds a set of tests and a parameterized test that will check the
behavior of the JUnit3 tests against standard JUnit3 code, e.g.
TestSuite(Class<?>) constructor and also JUnit3/JUnit4 bridging
code like JUnit38ClassRunner and SuiteMethod/Builder.