Refactor TestSuite code to make it reusable.
Separates the task of building a TestSuite (essentially a list
of Test objects) from a TestCase into two parts. The scanning of
the TestCase to identify the test methods and the construction
of a representation to run those tests. This allows the scanning
code to be reused to construct both a legacy JUnit3
representation based on TestSuite and a list of TestCase objects
and also a JUnit4 Runner based representation.

A new StandardTestCaseBuilder is added which uses the Runner
based representation created for a TestCase to run the test. It
is used by JUnit3Builder which will fallback to its previous
behavior using JUnit38ClassRunner if the StandardTestCaseBuilder
could not handle the supplied class. The reason it was not added
as a separate RunnerBuilder in AllDefaultPossibilitiesBuilder
was to minimise the impact on any extensions of that class that
may be relying on the current set of builders being the only
ones used.

The new Runner representations have a major advantage over the
standard JUnit3 ones which is that the TestCase instances have a
much shorter lifespan. When JUnit3 tests are run by
JUnit38ClassRunner the TestCase instances are created during
initialization phase and only discarded after all the tests have
been run. With a large number of tests that can require a very
large heap. When they are run using the new Runners the TestCase
instances are created immediately prior to the test being run
and discarded immediately after which can greatly reduce the
heap size required to run all the tests.