Pass RunnerParams through Parameterized
Parameterized will create a Runner instance for each set of data
using a ParametersRunnerFactory. This adds RunnerParams to
TestWithParameters so that it can be passed to the factory and
on to the Runner it creates.

The BlockJUnit4ClassRunnerWithParams passes the RunnerParams to
its super class BlockJUnit4ClassRunner so that test method
validation can be deferred for Parameterized tests too.