Only call method annotated with @Test once
If a method annotated with @Test returns a type which implements
both MethodRule and TestRule then the method is called twice and
two rules are created and used. That happens because the work to
find the TestRules and MethodRules was separate and so the
methods had to be scanned twice. Scanning the methods required
them to be invoked so they were invoked twice. If it created a
new object that was not equal to the previously created object
then they would both be added to the list and both would be run.

Instead of taking two passes over the set of annotated members,
once for TestRule and then again for MethodRule this does it in
a single pass. It adds the object created into the appropriate
list based on its type.

Unfortunately, because the methods getTestRules() and rules()
(which get the TestRules and MethodRules respectively) can be
overridden it was not possible to combine them into one
method. Instead the work has to be done separately, prior to
calling those methods, temporarily stored in a ThreadLocal so it
could be accessed from those methods without changing their
signature. The methods were called to pick up any changes made
by overridding methods and then the results were applied.

A ThreadLocal was used instead of instance fields in the
BlockJUnitClassRunner because those instance fields would need
to be accessed within a synchronized block to protect against
problems when running the test methods in parallel. That would
mean that the calls to getTestRules() and rules() would need to
be made from within the synchronized block which could cause
problems.