Add TargetedTestRule
There are currently two different types of rules, TestRule and
MethodRule that each have different capabilities to the other.
The former is generic and not coupled to a particular Runner
structure but does not have access to the object being tested.
It can therefore be used with both @ClassRule and @Rule
annotations. The latter is coupled to Runners that run
FrameworkMethod instances but has access to the test object so
can do more, e.g. inject values into it. As it has access to the
test object it can't be used with @ClassRule annotations.

TargetedTestRule combines parts of both. It is generic like
TestRule so could be applied globally and has access to the test
object like MethodRule.

The main difference between TargetedTestRule and MethodRule is
that they use Description and FrameworkMethod respectively to
identify the test. They both have access to the 'method' name,
class name and annotations associated with the 'method' so for
the rule implementations they are for all practical purposes
equivalent. The FrameworkMethod does provide access to the
underlying java.lang.reflect.Method object but it's not clear
why a rule would need that.