Discard Test referenced by JUnit38ClassRunner after running it
One of the problems with JUnit 3 tests, even when run under
JUnit 4 is that the lifespan of the test objects is the whole
test run. This discards the test immediately after the test has
run. As the test is used to generate the Description the current
Description is saved away before discarding the Test.